scripts/audio/requirements-azure.txt
# Azure ML audio transcription requirements
# GPU-enabled packages for NVIDIA T4

# Core ML packages
# Updated to 2.4.1 for vox-profile compatibility
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1

# WhisperX and dependencies
# Using latest whisperx (3.1.5+) for compatibility with faster-whisper 1.0+
whisperx
openai-whisper
faster-whisper
transformers>=4.46.3
pyannote.audio
huggingface_hub[hf_xet]

# Emotion analysis
# vox-profile for dimensional emotion analysis (arousal, valence, dominance)
git+https://github.com/tiantiaf0627/vox-profile-release.git

# Additional dependencies
pandas
numpy<2
scipy
librosa
soundfile
ffmpeg-python
tomlkit

# Azure ML SDK
azure-ai-ml
azure-storage-blob
azure-identity
